Staging

+++ app/Http/Controllers/Dashboard/SeminarPenelitianController.php
@@ -13,25 +13,26 @@ class SeminarPenelitianController extends Controller
 {
     public function index()
     {
-        if (Auth::user()->dosen && Auth::user()->dosen->dosen_pengujis->pluck('id')) {
-            $dosenPenguji = auth()->user()->dosen->dosen_pengujis;
-            // get id dosen penguji yang login
-            $dosenPengujiId = $dosenPenguji->pluck('id');
+        if (Auth::user()->dosen && Auth::user()->dosen->dosen_penguji) {
+            $dosenPengujiId = auth()->user()->dosen->dosen_penguji->id;
+            $seminarPenelitianId = SeminarPenelitianNilai::where('dosen_penguji_id', $dosenPengujiId)->pluck('seminar_penelitian_id'); // Ambil daftar seminar penelitian yang terkait dengan dosen penguji
+            $seminarPenelitians = SeminarPenelitian::whereIn('id', $seminarPenelitianId)->orderByDesc('created_at')->get(); // Tampilkan daftar seminar penelitian yang terkait
 
-            // Ambil daftar seminar penelitian yang terkait dengan dosen penguji
-            $seminarPenelitianId = SeminarPenelitianNilai::whereIn('dosen_penguji_id', $dosenPengujiId)->pluck('seminar_penelitian_id');
-
-            // Tampilkan daftar seminar penelitian yang terkait
-            $seminarPenelitians = SeminarPenelitian::whereIn('id', $seminarPenelitianId)->get();
+            $pengujiBelumNilai = SeminarPenelitianNilai::where('dosen_penguji_id', $dosenPengujiId)->whereNull('nilai')->pluck('seminar_penelitian_id'); // Ambil dosen penguji yang belum memberikan nilai
 
             return view('pages.dashboard.seminar_penelitian.index', [
                 'seminarPenelitians' => $seminarPenelitians,
+                'dosenPengujiId' => $dosenPengujiId,
+                'pengujiBelumNilai' => $pengujiBelumNilai
             ]);
         } else {
-            $seminarPenelitians = SeminarPenelitian::all();
+            $seminarPenelitians = SeminarPenelitian::orderByDesc('created_at')->get();
+            $pengujiBelumNilai = SeminarPenelitianNilai::whereNull('nilai')->pluck('seminar_penelitian_id');
 
             return view('pages.dashboard.seminar_penelitian.index', [
-                'seminarPenelitians' => SeminarPenelitian::all(),
+                'seminarPenelitians' => $seminarPenelitians,
+                'pengujiBelumNilai' => $pengujiBelumNilai
+
             ]);
         }
     }
@@ -40,7 +41,13 @@ class SeminarPenelitianController extends Controller
     {
         $admin = Auth::user()->role->nama == 'admin';
         if ($admin) {
-            $dosenSeminarPenelitians = DosenPenguji::with('dosen.user')->get();
+            $dosenSeminarPenelitians = DosenPenguji::with('dosen')
+                                                    ->join('dosens', 'dosen_pengujis.dosen_id', '=', 'dosens.id')
+                                                    ->join('users', 'dosens.user_id', '=', 'users.id')
+                                                    ->select(
+                                                        'dosen_pengujis.id as id',
+                                                        'users.nama as nama'
+                                                    )->orderBy('users.nama')->get(); //mengambil daftar dosen seminar penelitian
             $selectedDosenPenelitian = $seminarPenelitian->seminar_penelitian_nilais->pluck('dosen_penguji_id')->all();
 
             return view('pages.dashboard.seminar_penelitian.show', [
@@ -49,8 +56,14 @@ class SeminarPenelitianController extends Controller
                 'selectedDosenPenelitian' => $selectedDosenPenelitian
             ]);
         } else {
-            $dosenSeminarPenelitians = $seminarPenelitian->seminar_penelitian_nilais()->with('dosen_penguji.dosen.user')->get();
-            $selectedDosenPenelitian = $seminarPenelitian->seminar_penelitian_nilais()->pluck('dosen_penguji_id')->all();
+            $dosenSeminarPenelitians = $seminarPenelitian->seminar_penelitian_nilais()->with('dosen_penguji.dosen.user')
+                                                                                        ->join('dosen_pengujis', 'seminar_penelitian_nilai.dosen_penguji_id', '=', 'dosen_pengujis.id')
+                                                                                        ->join('dosens', 'dosen_pengujis.dosen_id', '=', 'dosens.id')
+                                                                                        ->join('users', 'dosens.user_id', '=', 'users.id')
+                                                                                        ->where('seminar_penelitian_nilai.seminar_penelitian_id', $seminarPenelitian->id)
+                                                                                        ->orderBy('users.nama')
+                                                                                        ->get(); //mengambil daftar dosen seminar proposal yang terkait dengan seminar proposal
+            $selectedDosenPenelitian = $seminarPenelitian->seminar_penelitian_nilais->pluck('dosen_penguji_id')->all();
 
             return view('pages.dashboard.seminar_penelitian.show', [
                 'seminarPenelitian' => $seminarPenelitian,
@@ -62,18 +75,24 @@ class SeminarPenelitianController extends Controller
 
     public function create()
     {
-        $mahasiswa = auth()->user()->mahasiswa;
-        $tugasAkhir = $mahasiswa->tugas_akhir;
-        $seminarProposal = $tugasAkhir->seminar_proposal;
+        $tugasAkhirMahasiswa = Auth::user()->mahasiswa->tugas_akhir;
+        $seminarProposal = $tugasAkhirMahasiswa->seminar_proposal;
 
-        // periksa nilai seminar proposal
-        if (!isset($seminarProposal->nilai_akhir)) {
-            return redirect()->back()->with('error', 'Mohon Maaf, Harap lengkapi penilaian Seminar Proposal Anda');
-        }
-        elseif ($seminarProposal->nilai_akhir) {
-            return view('pages.dashboard.seminar_penelitian.create', [
-                'tugasAkhir' => $tugasAkhir
-            ]);
+        if (isset($seminarProposal->nilai_akhir)) {
+            $dosenPengujiBelumNilai = $seminarProposal->seminar_proposal_nilais()->whereNull('nilai')->with('dosen_penguji')->get();
+
+
+            if ($dosenPengujiBelumNilai->isEmpty()) {
+                return view('pages.dashboard.seminar_penelitian.create', [
+                    'tugasAkhir' => $tugasAkhirMahasiswa
+                ]);
+            } else {
+                session()->flash('error', 'Mohon Maaf, masih terdapat Dosen Penguji yang belum menilai Seminar Proposal anda');
+                return redirect()->back();
+            }
+        } else {
+            session()->flash('error', 'Mohon Maaf, anda belum memiliki Seminar Proposal');
+            return redirect()->back();
         }
     }
 
@@ -82,37 +101,58 @@ class SeminarPenelitianController extends Controller
         $request->validate([
             'tempat' => 'nullable',
             'tanggal' => 'nullable',
-            'waktu' => 'nullable',
+            'waktu_mulai' => 'nullable',
+            'waktu_selesai' => 'nullable',
             'tugas_akhir_id' => 'required|exists:tugas_akhirs,id'
         ]);
 
         $tugasAkhir = TugasAkhir::find($request->tugas_akhir_id);
         if ($tugasAkhir->seminar_penelitian) {
-            return redirect()->back()->with('error', 'Mohon Maaf, Anda sudah menambahkan sidang penelitian');
+            session()->flash('error', 'Mohon Maaf, anda sudah menambahkan Seminar Penelitian');
+            return redirect()->back();
         }
 
         $seminarPenelitian = SeminarPenelitian::create([
             'tugas_akhir_id' => $request->tugas_akhir_id
         ]);
 
-        return redirect()->route('seminar-penelitian.show', ['seminarPenelitian' => $seminarPenelitian->id])->with('success', 'Seminar Penelitian berhasil ditambahkan.');
+        session()->flash('success', 'Seminar Penelitian berhasil ditambahkan');
+        return redirect()->route('seminar-penelitian.show', ['seminarPenelitian' => $seminarPenelitian->id]);
     }
 
     public function update(Request $request, SeminarPenelitian $seminarPenelitian)
     {
         $validate = $request->validate([
             'tanggal' => 'required|date',
-            'waktu' => 'required',
+            'waktu_mulai' => 'required',
+            'waktu_selesai' => 'required',
             'tempat' => 'required',
         ]);
 
+        $tanggal = $request->tanggal;
+        $tempat = $request->tempat;
+
+        $bentrok = SeminarPenelitian::where('id', '!=', $seminarPenelitian->id)
+                    ->where('tempat', $tempat)
+                    ->where('tanggal', $tanggal)
+                    ->where(function($query) use ($validate){
+                        $query->where(function ($query) use ($validate) {
+                            $query->whereBetween('waktu_mulai', [$validate['waktu_mulai'], $validate['waktu_selesai']])
+                                ->orWhereBetween('waktu_selesai', [$validate['waktu_mulai'], $validate['waktu_selesai']]);
+                        })
+                        ->orWhere(function ($query) use ($validate) {
+                            $query->where('waktu_mulai', '<=', $validate['waktu_mulai'])
+                                ->where('waktu_selesai', '>=', $validate['waktu_selesai']);
+                        });
+                    })->exists();
+        if ($bentrok) {
+            return redirect()->back()->with('error', 'Maaf, terdapat bentrok dengan acara lain pada waktu dan tempat tersebut.');
+        }
+
         $seminarPenelitian->update($validate);
 
-        // Ambil daftar dosen penguji yang dipilih
-        $selectedDosenPengujiIds = $request->input('dosen_penguji_id', []);
-
-        // hapus data dosen penguji yang tidak dipilih pada tabel seminar_penelitian_nilai
-        $seminarPenelitian->seminar_penelitian_nilais()->whereNotIn('dosen_penguji_id', $selectedDosenPengujiIds)->delete();
+        $selectedDosenPengujiIds = $request->input('dosen_penguji_', []); // Ambil daftar dosen penguji yang dipilih
+        $seminarPenelitian->seminar_penelitian_nilais()->whereNotIn('dosen_penguji_id', $selectedDosenPengujiIds)->delete(); // hapus data dosen penguji yang tidak dipilih pada tabel seminar_penelitian_nilai
 
         // Tambahkan data dosen penguji terpilih ke seminar_penelitian_nilai
         foreach ($selectedDosenPengujiIds as $dosenPengujiId) {
@@ -123,6 +163,7 @@ class SeminarPenelitianController extends Controller
             $seminarPenelitianNilai->save();
         }
 
-        return redirect()->route('seminar-penelitian.show', ['seminarPenelitian' => $seminarPenelitian->id])->with('success', 'Data Seminar Proposal berhasil diperbarui.');
+        session()->flash('success', 'Data Seminar Proposal berhasil diperbarui');
+        return redirect()->route('seminar-penelitian.show', ['seminarPenelitian' => $seminarPenelitian->id]);
     }
 }
